bot/requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
pydantic==2.5.2
